+++ protocols/kad/src/kbucket.rs
@@ -75,7 +75,7 @@ mod key;
 pub use bucket::NodeStatus;
 pub use entry::*;
 
-use arrayvec::{self, ArrayVec};
+use arrayvec::ArrayVec;
 use bucket::KBucket;
 use std::collections::VecDeque;
 use std::time::{Duration, Instant};
@@ -344,7 +344,7 @@ enum ClosestBucketsIterState {
     /// The starting state of the iterator yields the first bucket index and
     /// then transitions to `ZoomIn`.
     Start(BucketIndex),
-    /// The iterator "zooms in" to to yield the next bucket cotaining nodes that
+    /// The iterator "zooms in" to yield the next bucket cotaining nodes that
     /// are incrementally closer to the local node but further from the `target`.
     /// These buckets are identified by a `1` in the corresponding bit position
     /// of the distance bit string. When bucket `0` is reached, the iterator